requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dateutil==2.8.2
ffmpeg-python==0.2.0
Pillow==10.1.0
transformers==4.35.2
torch==2.1.1